Changed base type of labelObj size, minsize and maxsize from int to double (#2766)
git-svn-id: http://svn.osgeo.org/mapserver/trunk@8482 7532c77e-422f-0410-93f4-f0b67bdd69e2

+++ mapimagemap.c
@@ -1848,7 +1848,7 @@ int msDrawLabelCacheIM(imageObj* img, mapObj *map)
       continue; /* not an error, just don't want to do anything */
 
     if(cachePtr->label.type == MS_TRUETYPE)
-      cachePtr->label.size = (int)(cachePtr->label.size*layerPtr->scalefactor);
+      cachePtr->label.size = (cachePtr->label.size*layerPtr->scalefactor);
 
     if(msGetLabelSize(img,cachePtr->text, labelPtr, &r, &(map->fontset), layerPtr->scalefactor, MS_TRUE,NULL) == -1)
       return(-1);
